add the all-in one analytical model for unequal power
